Split drivers.
Transport aware drivers are now in the client crate so that protocol
drivers can be used independently of the client transport.
