+++ client/Cargo.toml
@@ -7,8 +7,19 @@ keywords = ["mpc", "tss", "relay", "crypto", "e2ee"]
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/mpc-sdk/framework"
 
+[features]
+default = ["protocols", "signers"]
+cggmp = ["mpc-driver/cggmp"]
+ecdsa = ["mpc-driver/ecdsa"]
+eddsa = ["mpc-driver/eddsa"]
+schnorr = ["mpc-driver/schnorr"]
+frost-ed25519 = ["mpc-driver/frost-ed25519"]
+signers = ["ecdsa", "eddsa", "schnorr"]
+protocols = ["cggmp", "frost-ed25519"]
+
 [dependencies]
 mpc-protocol = { version = "0.7", path = "../protocol", features = ["zlib"] }
+mpc-driver = { version = "0.7", path = "../driver" }
 thiserror = "1"
 serde = { version = "1", features = ["derive"] }
 tracing = "0.1"
