+++ client/src/error.rs
@@ -54,6 +54,25 @@ pub enum Error {
     #[error("web socket failed to send")]
     WebSocketSend,
 
+    /// Error generated when meeting identifiers are not unique.
+    #[error("meeting identifiers must be unique")]
+    MeetingIdentifiersNotUnique,
+
+    /// Error generated when the meeting initiator does not exist
+    /// in the list of meeting point identifiers.
+    #[error("meeting initiator must exist in list of identifiers")]
+    MeetingInitiatorNotExist,
+
+    /// Error generated when the user's public key is not in the
+    /// list of session participants.
+    #[error("public key {0} is not a session participant")]
+    NotSessionParticipant(String),
+
+    #[cfg(feature = "cggmp")]
+    /// Could not locate ack for key init phase.
+    #[error("could not find an ACK for key init phase")]
+    NoKeyInitAck,
+
     /// Javascript string error message.
     #[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
     #[error("{0}")]
@@ -74,6 +93,12 @@ pub enum Error {
     #[error("stream and sink reunite failed")]
     StreamReunite,
 
+    /// Generic boxed error.
+    #[error(transparent)]
+    Generic(
+        #[from] Box<dyn std::error::Error + Send + Sync + 'static>,
+    ),
+
     /// Error generated by input/output.
     #[error(transparent)]
     Io(#[from] std::io::Error),
@@ -81,6 +106,10 @@ pub enum Error {
     /// Error generated by the protocol library.
     #[error(transparent)]
     Protocol(#[from] mpc_protocol::Error),
+
+    /// Error generated by the driver library.
+    #[error(transparent)]
+    Driver(#[from] mpc_driver::Error),
 
     /// Error generated by the noise protocol library.
     #[error(transparent)]
@@ -102,6 +131,21 @@ pub enum Error {
     ResponseMpscSend(
         #[from] tokio::sync::mpsc::error::SendError<ResponseMessage>,
     ),
+
+    #[cfg(feature = "frost-ed25519")]
+    /// FROST library error.
+    #[error(transparent)]
+    FrostEd25519Core(#[from] mpc_driver::frost_ed25519::Error),
+
+    #[cfg(feature = "cggmp")]
+    /// CGGMP library error.
+    #[error(transparent)]
+    Cggmp(#[from] mpc_driver::cggmp::Error),
+
+    #[cfg(feature = "frost-ed25519")]
+    /// FROST library error.
+    #[error(transparent)]
+    FrostEd25519(#[from] mpc_driver::frost::Error),
 }
 
 #[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
