+++ client/src/protocols/frost/mod.rs
@@ -1,0 +1,4 @@
+//! FROST protocol implementations.
+#[cfg(feature = "frost-ed25519")]
+pub mod ed25519;
+
