+++ driver/Cargo.toml
@@ -19,7 +19,7 @@ full = ["protocols", "signers"]
 
 [dependencies]
 mpc-protocol = { version = "0.7", path = "../protocol" }
-mpc-client = { version = "0.7", path = "../client" }
+# mpc-client = { version = "0.7", path = "../client" }
 thiserror = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
