+++ driver/src/cggmp/error.rs
@@ -31,17 +31,9 @@ pub enum Error {
     #[error("protocol is not finished, another round is available")]
     NotFinished,
 
-    /// Error generated by the client library.
-    #[error(transparent)]
-    Client(#[from] mpc_client::Error),
-
     /// Protocol library errors.
     #[error(transparent)]
     Protocol(#[from] mpc_protocol::Error),
-
-    /// Driver library error.
-    #[error(transparent)]
-    Driver(#[from] Box<crate::Error>),
 
     /// Error generated converting integers.
     #[error(transparent)]
