+++ driver/src/cggmp/key_gen.rs
@@ -1,7 +1,4 @@
 //! Key generation for CGGMP.
-use async_trait::async_trait;
-use mpc_client::{NetworkTransport, Transport};
-use mpc_protocol::{hex, Event, SessionState};
 use rand::rngs::OsRng;
 use std::collections::BTreeSet;
 
@@ -16,88 +13,12 @@ use synedrion::{
     AuxInfo, KeyGenResult, KeyShare, SchemeParams, SessionId,
 };
 
-use crate::{
-    key_to_str, Bridge, Driver, ProtocolDriver, RoundInfo, RoundMsg,
-};
+use crate::{ProtocolDriver, RoundInfo, RoundMsg};
 
 use super::MessageOut;
 
-/// CGGMP key generation driver.
-pub struct KeyGenDriver<P>
-where
-    P: SchemeParams + 'static,
-{
-    bridge: Bridge<CggmpDriver<P>>,
-}
-
-impl<P> KeyGenDriver<P>
-where
-    P: SchemeParams + 'static,
-{
-    /// Create a new CGGMP key generator.
-    pub fn new(
-        transport: Transport,
-        session: SessionState,
-        session_id: SessionId,
-        signer: SigningKey,
-        verifiers: Vec<VerifyingKey>,
-    ) -> Result<Self> {
-        let party_number = session
-            .party_number(transport.public_key())
-            .ok_or_else(|| {
-                Error::NotSessionParticipant(hex::encode(
-                    transport.public_key(),
-                ))
-            })?;
-
-        let driver = CggmpDriver::new(session_id, signer, verifiers)?;
-
-        let bridge = Bridge {
-            transport,
-            driver: Some(driver),
-            session,
-            party_number,
-        };
-        Ok(Self { bridge })
-    }
-}
-
-#[async_trait]
-impl<P> Driver for KeyGenDriver<P>
-where
-    P: SchemeParams + 'static,
-{
-    type Error = Error;
-    type Output =
-        (KeyShare<P, VerifyingKey>, AuxInfo<P, VerifyingKey>);
-
-    async fn handle_event(
-        &mut self,
-        event: Event,
-    ) -> Result<Option<Self::Output>> {
-        self.bridge.handle_event(event).await
-    }
-
-    async fn execute(&mut self) -> Result<()> {
-        self.bridge.execute().await
-    }
-
-    fn into_transport(self) -> Transport {
-        self.bridge.transport
-    }
-}
-
-impl<P> From<KeyGenDriver<P>> for Transport
-where
-    P: SchemeParams + 'static,
-{
-    fn from(value: KeyGenDriver<P>) -> Self {
-        value.bridge.transport
-    }
-}
-
 /// CGGMP keygen driver.
-struct CggmpDriver<P>
+pub struct KeyGenDriver<P>
 where
     P: SchemeParams + 'static,
 {
@@ -116,7 +37,7 @@ where
     verifiers: Vec<VerifyingKey>,
 }
 
-impl<P> CggmpDriver<P>
+impl<P> KeyGenDriver<P>
 where
     P: SchemeParams + 'static,
 {
@@ -150,7 +71,7 @@ where
     }
 }
 
-impl<P> ProtocolDriver for CggmpDriver<P>
+impl<P> ProtocolDriver for KeyGenDriver<P>
 where
     P: SchemeParams + 'static,
 {
@@ -190,8 +111,10 @@ where
         let session = self.session.take().unwrap();
         let accum = self.accum.take().unwrap();
 
+        /*
         let key_str = key_to_str(&session.verifier());
         println!("{key_str}: finalizing the round");
+        */
 
         match session.finalize_round(&mut OsRng, accum).unwrap() {
             FinalizeOutcome::Success(result) => Ok(Some(result)),
