+++ driver/src/cggmp/key_refresh.rs
@@ -1,7 +1,4 @@
 //! Key refresh for CGGMP.
-use async_trait::async_trait;
-use mpc_client::{NetworkTransport, Transport};
-use mpc_protocol::{hex, Event, SessionState};
 use rand::rngs::OsRng;
 use std::collections::BTreeSet;
 
@@ -17,88 +14,12 @@ use synedrion::{
     SessionId,
 };
 
-use crate::{
-    key_to_str, Bridge, Driver, ProtocolDriver, RoundInfo, RoundMsg,
-};
+use crate::{ProtocolDriver, RoundInfo, RoundMsg};
 
 use super::MessageOut;
 
-/// CGGMP key refresh driver.
-pub struct KeyRefreshDriver<P>
-where
-    P: SchemeParams + 'static,
-{
-    bridge: Bridge<CggmpDriver<P>>,
-}
-
-impl<P> KeyRefreshDriver<P>
-where
-    P: SchemeParams + 'static,
-{
-    /// Create a new CGGMP refresh.
-    pub fn new(
-        transport: Transport,
-        session: SessionState,
-        session_id: SessionId,
-        signer: SigningKey,
-        verifiers: Vec<VerifyingKey>,
-    ) -> Result<Self> {
-        let party_number = session
-            .party_number(transport.public_key())
-            .ok_or_else(|| {
-                Error::NotSessionParticipant(hex::encode(
-                    transport.public_key(),
-                ))
-            })?;
-
-        let driver = CggmpDriver::new(session_id, signer, verifiers)?;
-
-        let bridge = Bridge {
-            transport,
-            driver: Some(driver),
-            session,
-            party_number,
-        };
-        Ok(Self { bridge })
-    }
-}
-
-#[async_trait]
-impl<P> Driver for KeyRefreshDriver<P>
-where
-    P: SchemeParams + 'static,
-{
-    type Error = Error;
-    type Output =
-        (KeyShareChange<P, VerifyingKey>, AuxInfo<P, VerifyingKey>);
-
-    async fn handle_event(
-        &mut self,
-        event: Event,
-    ) -> Result<Option<Self::Output>> {
-        self.bridge.handle_event(event).await
-    }
-
-    async fn execute(&mut self) -> Result<()> {
-        self.bridge.execute().await
-    }
-
-    fn into_transport(self) -> Transport {
-        self.bridge.transport
-    }
-}
-
-impl<P> From<KeyRefreshDriver<P>> for Transport
-where
-    P: SchemeParams + 'static,
-{
-    fn from(value: KeyRefreshDriver<P>) -> Self {
-        value.bridge.transport
-    }
-}
-
 /// CGGMP keygen driver.
-struct CggmpDriver<P>
+pub struct KeyRefreshDriver<P>
 where
     P: SchemeParams + 'static,
 {
@@ -117,7 +38,7 @@ where
     verifiers: Vec<VerifyingKey>,
 }
 
-impl<P> CggmpDriver<P>
+impl<P> KeyRefreshDriver<P>
 where
     P: SchemeParams + 'static,
 {
@@ -152,7 +73,7 @@ where
     }
 }
 
-impl<P> ProtocolDriver for CggmpDriver<P>
+impl<P> ProtocolDriver for KeyRefreshDriver<P>
 where
     P: SchemeParams + 'static,
 {
@@ -192,8 +113,10 @@ where
         let session = self.session.take().unwrap();
         let accum = self.accum.take().unwrap();
 
+        /*
         let key_str = key_to_str(&session.verifier());
         println!("{key_str}: finalizing the round");
+        */
 
         match session.finalize_round(&mut OsRng, accum).unwrap() {
             FinalizeOutcome::Success(result) => Ok(Some(result)),
