+++ driver/src/frost/error.rs
@@ -36,17 +36,9 @@ pub enum Error {
     #[error("attempt to proceed to round 3 without round 2 data")]
     Round3TooEarly,
 
-    /// Error generated by the client library.
-    #[error(transparent)]
-    Client(#[from] mpc_client::Error),
-
     /// Protocol library errors.
     #[error(transparent)]
     Protocol(#[from] mpc_protocol::Error),
-
-    /// Driver library error.
-    #[error(transparent)]
-    Driver(#[from] Box<crate::Error>),
 
     /// FROST library error.
     #[error(transparent)]
