+++ driver/src/frost/mod.rs
@@ -4,9 +4,6 @@ mod error;
 #[cfg(feature = "frost-ed25519")]
 pub mod ed25519;
 
-#[cfg(feature = "frost-ed25519")]
-pub use frost_ed25519;
-
 pub use error::Error;
 
 /// Result type for the FROST protocol.
